Tests/reorganize: move addStringAttachment tests to own file
