+++ test/PHPMailer/AddStringAttachmentTest.php
@@ -1,0 +1,54 @@
+<?php
+
+/**
+ * PHPMailer - PHP email transport unit tests.
+ * PHP version 5.5.
+ *
+ * @author    Marcus Bointon <phpmailer@synchromedia.co.uk>
+ * @author    Andy Prevost
+ * @copyright 2012 - 2020 Marcus Bointon
+ * @copyright 2004 - 2009 Andy Prevost
+ * @license   http://www.gnu.org/copyleft/lesser.html GNU Lesser General Public License
+ */
+
+namespace PHPMailer\Test\PHPMailer;
+
+use PHPMailer\PHPMailer\Exception;
+use PHPMailer\PHPMailer\PHPMailer;
+use PHPMailer\Test\SendTestCase;
+
+/**
+ * Test adding string attachments functionality.
+ */
+final class AddStringAttachmentTest extends SendTestCase
+{
+
+    /**
+     * Simple plain string attachment test.
+     */
+    public function testPlainStringAttachment()
+    {
+        $this->Mail->Body = 'Here is the text body';
+        $this->Mail->Subject .= ': Plain + StringAttachment';
+
+        $sAttachment = 'These characters are the content of the ' .
+            "string attachment.\nThis might be taken from a " .
+            'database or some other such thing. ';
+
+        $this->Mail->addStringAttachment($sAttachment, 'string_attach.txt');
+
+        $this->buildBody();
+        self::assertTrue($this->Mail->send(), $this->Mail->ErrorInfo);
+    }
+
+    /**
+     * Expect exceptions on bad encoding
+     */
+    public function testStringAttachmentEncodingException()
+    {
+        $this->expectException(Exception::class);
+
+        $mail = new PHPMailer(true);
+        $mail->addStringAttachment('hello', 'test.txt', 'invalidencoding');
+    }
+}
